+++ src/activities/network/NetworkModeSelectionActivity.cpp
@@ -12,13 +12,15 @@
 namespace fui = freeink::ui;
 
 namespace {
-constexpr int MENU_ITEM_COUNT = 4;
+constexpr int MENU_ITEM_COUNT = 5;
 constexpr fui::ActionId ACTION_ROW = 1;
 constexpr StrId menuItems[MENU_ITEM_COUNT] = {StrId::STR_JOIN_NETWORK, StrId::STR_CALIBRE_WIRELESS,
-                                              StrId::STR_CREATE_HOTSPOT, StrId::STR_NEARBY_STATS_SYNC};
+                                              StrId::STR_CREATE_HOTSPOT, StrId::STR_RECEIVE_NEARBY_BOOK,
+                                              StrId::STR_NEARBY_STATS_SYNC};
 constexpr StrId menuDescs[MENU_ITEM_COUNT] = {StrId::STR_JOIN_DESC, StrId::STR_CALIBRE_DESC, StrId::STR_HOTSPOT_DESC,
-                                              StrId::STR_NEARBY_STATS_SYNC_DESC};
-constexpr UIIcon menuIcons[MENU_ITEM_COUNT] = {UIIcon::Wifi, UIIcon::Library, UIIcon::Hotspot, UIIcon::Transfer};
+                                              StrId::STR_RECEIVE_NEARBY_BOOK_DESC, StrId::STR_NEARBY_STATS_SYNC_DESC};
+constexpr UIIcon menuIcons[MENU_ITEM_COUNT] = {UIIcon::Wifi, UIIcon::Library, UIIcon::Hotspot, UIIcon::Transfer,
+                                               UIIcon::Transfer};
 }  // namespace
 
 NetworkModeSelectionActivity::NetworkModeSelectionActivity(GfxRenderer& renderer, MappedInputManager& mappedInput)
@@ -33,6 +35,8 @@ void NetworkModeSelectionActivity::selectCurrent() {
   } else if (selectedIndex == 2) {
     mode = NetworkMode::CREATE_HOTSPOT;
   } else if (selectedIndex == 3) {
+    mode = NetworkMode::NEARBY_BOOK_RECEIVE;
+  } else if (selectedIndex == 4) {
     mode = NetworkMode::NEARBY_STATS_SYNC;
   }
   onModeSelected(mode);