+++ src/activities/network/NetworkModeSelectionActivity.h
@@ -9,7 +9,7 @@
 #include "activities/Activity.h"
 #include "util/ButtonNavigator.h"
 
-enum class NetworkMode { JOIN_NETWORK, CONNECT_CALIBRE, CREATE_HOTSPOT, NEARBY_STATS_SYNC };
+enum class NetworkMode { JOIN_NETWORK, CONNECT_CALIBRE, CREATE_HOTSPOT, NEARBY_BOOK_RECEIVE, NEARBY_STATS_SYNC };
 
 /**
  * NetworkModeSelectionActivity presents the user with a choice:
@@ -17,6 +17,7 @@ enum class NetworkMode { JOIN_NETWORK, CONNECT_CALIBRE, CREATE_HOTSPOT, NEARBY_S
  * - "Connect to Calibre" - Use Calibre wireless device transfers
  * - "Create Hotspot" - Create an Access Point that others can connect to (AP mode)
  * - "Nearby Stats Sync" - Sync reading stats directly with a nearby reader
+ * - "Receive Nearby Book" - Receive a book directly from another reader
  *
  * The onModeSelected callback is called with the user's choice.
  * The onCancel callback is called if the user presses back.